requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
backcall==0.1.0
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
cryptography==2.4.2
decorator==4.3.0
defusedxml==0.5.0
dnspython==1.15.0
exchangelib==1.12.0
Flask-Cors==3.0.7
future==0.17.1
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jira==2.0.0
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.1.1
marshmallow==2.19.1
mccabe==0.6.1
ntlm-auth==1.2.0
oauthlib==2.1.0
parso==0.3.1
pbr==5.1.1
pickleshare==0.7.5
prompt-toolkit==2.0.7
psutil==5.4.8
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.1
pyserial==3.4
python-dateutil==2.7.5
pytz==2018.7
requests==2.19.1
requests-ntlm==1.1.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
six==1.11.0
SQLAlchemy==1.3.3
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.15.0
wrapt==1.10.11
